// src/Models/Sector/sector.ts
export class Sector{
    SrNo: number;
    Result: number;
    SectorId: any;
    SectorCode: string;
    SectorName: string;
    CountryCode: string;
    CountryName: string;
    CreatedBy: string;
    CreatedOn: string;
    UpdatedBy: string;
    UpdatedOn: string;
}

export class Country{
    CountryCode: string;
    CountryName: string;
}